add full search space
